CuteEyeRobot/pyproject.toml
[project]
name = "cute-eye-robot-tools"
version = "0.1.0"
description = "Serial test tools for CuteEyeRobot"
requires-python = ">=3.10"
dependencies = [
  "pyserial>=3.5",
]

[tool.uv]
package = false
